+++ extensions/easydict/src/dictionary/linguee/linguee.ts
@@ -2,7 +2,7 @@
  * @author: tisfeng
  * @createTime: 2022-07-24 17:58
  * @lastEditor: tisfeng
- * @lastEditTime: 2022-09-02 23:19
+ * @lastEditTime: 2022-09-11 17:13
  * @fileName: linguee.ts
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.
@@ -80,7 +80,6 @@ export async function rquestLingueeDictionary(queryWordInfo: QueryWordInfo): Pro
             toLanguage: queryWordInfo.toLanguage,
           };
         }
-
         resolve(lingueeTypeResult);
       })
       .catch((error: AxiosError) => {