+++ extensions/easydict/src/types.ts
@@ -2,7 +2,7 @@
  * @author: tisfeng
  * @createTime: 2022-06-04 21:58
  * @lastEditor: tisfeng
- * @lastEditTime: 2022-09-02 13:11
+ * @lastEditTime: 2022-09-11 17:44
  * @fileName: types.ts
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.
@@ -93,16 +93,11 @@ export interface TencentTranslateResult extends TextTranslateResponse {
   Error: TencentError;
 }
 
+// {"Code":"InvalidParameterValue","Message":"不支持的语种：ar_to_zh"}
 export interface TencentError {
   Code: string;
   Message: string;
 }
-// {"Code":"InvalidParameterValue","Message":"不支持的语种：ar_to_zh"}
-
-// export interface TencentTranslateResult {
-//   Response: TencentTranslateResponse;
-// }
-// export type TencentTranslateResponse = TextTranslateResponse;
 
 export interface CaiyunTranslateResult {
   rc: string;