+++ extensions/easydict/src/utils.ts
@@ -2,13 +2,14 @@
  * @author: tisfeng
  * @createTime: 2022-08-04 12:28
  * @lastEditor: tisfeng
- * @lastEditTime: 2022-09-02 23:11
+ * @lastEditTime: 2022-09-05 23:14
  * @fileName: utils.ts
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.
  */
 
 import { Clipboard, getApplications, LocalStorage, showToast, Toast } from "@raycast/api";
+import CryptoJS from "crypto-js";
 import { AxiosError } from "axios";
 import { clipboardQueryTextKey } from "./consts";
 import { LanguageDetectType } from "./detectLanauge/types";
@@ -18,7 +19,7 @@ import { myPreferences } from "./preferences";
 import { Easydict } from "./releaseVersion/versionInfo";
 import {
   DicionaryType,
-  ListItemDisplayType,
+  ListDisplayItem,
   QueryRecoredItem,
   QueryType,
   RequestErrorInfo,
@@ -227,8 +228,12 @@ export function checkIsLanguageDetectType(type: RequestType): boolean {
 /**
  * check type is YoudaoDictionaryListItem type.
  */
-export function checkIsYoudaoDictionaryListItemType(type: ListItemDisplayType): boolean {
-  if (Object.values(YoudaoDictionaryListItemType).includes(type as YoudaoDictionaryListItemType)) {
+export function checkIsYoudaoDictionaryListItem(listItem: ListDisplayItem): boolean {
+  const { queryType, displayType } = listItem;
+  if (
+    queryType === DicionaryType.Youdao &&
+    Object.values(YoudaoDictionaryListItemType).includes(displayType as YoudaoDictionaryListItemType)
+  ) {
     return true;
   }
   return false;
@@ -237,9 +242,17 @@ export function checkIsYoudaoDictionaryListItemType(type: ListItemDisplayType): 
 /**
  * check type is LingueeListItem type.
  */
-export function checkIsLingueeListItemType(type: ListItemDisplayType): boolean {
-  if (Object.values(LingueeListItemType).includes(type as LingueeListItemType)) {
+export function checkIsLingueeListItem(listItem: ListDisplayItem): boolean {
+  const { queryType, displayType } = listItem;
+  if (
+    queryType === DicionaryType.Linguee &&
+    Object.values(LingueeListItemType).includes(displayType as LingueeListItemType)
+  ) {
     return true;
   }
   return false;
+}
+
+export function md5(text: string): string {
+  return CryptoJS.MD5(text).toString();
 }